requirement.txt
backcall==0.1.0
Click==7.0
decorator==4.4.0
Flask==1.0.3
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.5
hurry.filesize==0.9
Flask-RESTful==0.3.8
